Split mainloopschedulers package into mainloop and eventloop
