+++ docs/reference_scheduler.rst
@@ -8,8 +8,10 @@ Schedulers
                 HistoricalScheduler, ImmediateScheduler, NewThreadScheduler,
                 ThreadPoolScheduler, TimeoutScheduler, VirtualTimeScheduler
 
-.. automodule:: rx.concurrency.mainloopscheduler
+.. automodule:: rx.scheduler.eventloop
     :members: AsyncIOScheduler, AsyncIOThreadSafeScheduler, EventletScheduler,
-                GEventScheduler, GtkScheduler, IOLoopScheduler,
-                PyGameScheduler, QtScheduler, TkinterScheduler
-                TwistedScheduler, WxScheduler
+                GEventScheduler, IOLoopScheduler, TwistedScheduler
+
+.. automodule:: rx.scheduler.mainloop
+    :members: GtkScheduler, PyGameScheduler, QtScheduler,
+                TkinterScheduler, WxScheduler
