+++ examples/timeflies/timeflies_gtk.py
@@ -1,7 +1,7 @@
 import rx
 from rx import operators as ops
 from rx.subjects import Subject
-from rx.scheduler.mainloopscheduler import GtkScheduler
+from rx.scheduler.mainloop import GtkScheduler
 
 import gi
 gi.require_version('Gtk', '3.0')
