+++ examples/timeflies/timeflies_tkinter.py
@@ -3,7 +3,7 @@ from tkinter import Tk, Label, Frame
 import rx
 from rx import operators as ops
 from rx.subjects import Subject
-from rx.scheduler.mainloopscheduler import TkinterScheduler
+from rx.scheduler.mainloop import TkinterScheduler
 
 
 def main():
