+++ rx/scheduler/__init__.py
@@ -1,11 +1,11 @@
 from .scheduleditem import ScheduledItem
 
-from .immediatescheduler import ImmediateScheduler, immediate_scheduler
+from .catchscheduler import CatchScheduler
 from .currentthreadscheduler import CurrentThreadScheduler, current_thread_scheduler
-from .virtualtimescheduler import VirtualTimeScheduler
-from .timeoutscheduler import TimeoutScheduler, timeout_scheduler
-from .newthreadscheduler import NewThreadScheduler
-from .threadpoolscheduler import ThreadPoolScheduler
 from .eventloopscheduler import EventLoopScheduler
 from .historicalscheduler import HistoricalScheduler
-from .catchscheduler import CatchScheduler
+from .immediatescheduler import ImmediateScheduler, immediate_scheduler
+from .newthreadscheduler import NewThreadScheduler
+from .threadpoolscheduler import ThreadPoolScheduler
+from .timeoutscheduler import TimeoutScheduler, timeout_scheduler
+from .virtualtimescheduler import VirtualTimeScheduler
