+++ rx/scheduler/mainloop/gtkscheduler.py
