--- rx/scheduler/mainloopscheduler/__init__.py
@@ -1,11 +1,0 @@
-from .asyncioscheduler import AsyncIOScheduler
-from .asynciothreadsafescheduler import AsyncIOThreadSafeScheduler
-from .ioloopscheduler import IOLoopScheduler
-from .geventscheduler import GEventScheduler
-from .gtkscheduler import GtkScheduler
-from .twistedscheduler import TwistedScheduler
-from .tkinterscheduler import TkinterScheduler
-from .pygamescheduler import PyGameScheduler
-from .qtscheduler import QtScheduler
-from .wxscheduler import WxScheduler
-from .eventletscheduler import EventletScheduler
