+++ setup.py
@@ -56,9 +56,9 @@ setup(
 
     package_data={'rx': ['py.typed']},
     packages=['rx', 'rx.internal', 'rx.core', 'rx.core.abc',
-              'rx.core.operators', 'rx.core.observable',
-              'rx.core.observer', 'rx.core.operators.connectable',
-              'rx.scheduler', 'rx.scheduler.mainloopscheduler',
+              'rx.core.operators', 'rx.core.operators.connectable',
+              'rx.core.observer', 'rx.core.observable',
+              'rx.scheduler', 'rx.scheduler.eventloop', 'rx.scheduler.mainloop',
               'rx.operators', 'rx.disposable', 'rx.subjects',
               'rx.testing'],
     package_dir={'rx': 'rx'},
