+++ tests/test_scheduler/test_eventloop/test_asynciothreadsafescheduler.py
@@ -4,7 +4,7 @@ import asyncio
 import threading
 from datetime import datetime, timedelta
 
-from rx.scheduler.mainloopscheduler import AsyncIOThreadSafeScheduler
+from rx.scheduler.eventloop import AsyncIOThreadSafeScheduler
 
 
 class TestAsyncIOThreadSafeScheduler(unittest.TestCase):
