+++ tests/test_scheduler/test_eventloop/test_eventletscheduler.py
@@ -4,7 +4,7 @@ import unittest
 from datetime import datetime, timedelta
 from time import sleep
 
-from rx.scheduler.mainloopscheduler import EventletScheduler
+from rx.scheduler.eventloop import EventletScheduler
 
 
 eventlet = pytest.importorskip("eventlet")
