+++ tests/test_scheduler/test_eventloop/test_geventscheduler.py
@@ -3,7 +3,7 @@ import unittest
 
 from datetime import datetime, timedelta
 
-from rx.scheduler.mainloopscheduler import GEventScheduler
+from rx.scheduler.eventloop import GEventScheduler
 
 
 gevent = pytest.importorskip("gevent")
