+++ tests/test_scheduler/test_eventloop/test_ioloopscheduler.py
@@ -7,7 +7,7 @@ from time import sleep
 tornado = pytest.importorskip("tornado")
 from tornado import ioloop
 
-from rx.scheduler.mainloopscheduler import IOLoopScheduler
+from rx.scheduler.eventloop import IOLoopScheduler
 
 
 class TestIOLoopScheduler(unittest.TestCase):
