+++ tests/test_scheduler/test_mainloop/test_pygamescheduler.py
@@ -4,7 +4,7 @@ import unittest
 from datetime import timedelta
 from time import sleep
 
-from rx.scheduler.mainloopscheduler import PyGameScheduler
+from rx.scheduler.mainloop import PyGameScheduler
 from rx.internal.basic import default_now
