+++ tests/test_scheduler/test_mainloop/test_qtscheduler_pyqt5.py
@@ -6,7 +6,7 @@ from time import sleep
 
 QtCore = pytest.importorskip('PyQt5.QtCore')
 
-from rx.scheduler.mainloopscheduler import QtScheduler
+from rx.scheduler.mainloop import QtScheduler
 from rx.internal.basic import default_now
